Notebooks: Review-readiness cleanup
Splits the block body out of the oversized editor component, consolidates the
shared time formatting helpers, fixes strict-typecheck errors at the generated
client type seams (element union access, resource annotations typing), adds
orchestration tests for the add-to-notebook flow, and documents private
preview review notes (permissions shortcut, security posture, FE/BE PR split,
coverage map) in the feature README.
